requirements.txt
fastapi
uvicorn
python-dotenv
pandas
sqlalchemy
psycopg2-binary
